Refactor game stats and display updates
Updated Board::set_legend to use int for loop variables. Enhanced Display methods and layouts to show score and time played. Modified Game::start to return const Game::Statistics& and added methods to save and get stats. Renamed Level and Mario methods for consistency. Included new headers and functions for timer management. Updated game layout files with minor changes.

+++ Board.cpp
@@ -55,8 +55,8 @@ void Board::set_legend() {
     Coordinates pos = map[LEGEND][0];
 
 	// Print the legend frame (fill it with AIR to make sure the legend is clear for data)
-    for (size_t i = 0; i < LEGEND_HEIGHT; i++) {
-        for (size_t j = 0; j < LEGEND_WIDTH; j++) {
+    for (int i = 0; i < LEGEND_HEIGHT; i++) {
+        for (int j = 0; j < LEGEND_WIDTH; j++) {
             if (i == 2 || j == 0 || j == 19) {
 				set_char(pos.x + j, pos.y + i, WALL);
 			}
